Add auth checks in some event service methods

+++ MeetMingler/MeetMingler.BLL/Services/Implementations/EventService.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using System.Linq.Expressions;
+using AutoMapper;
 using MeetMingler.BLL.Filters;
 using MeetMingler.BLL.Models;
 using MeetMingler.BLL.Models.Event;
@@ -16,15 +17,15 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
         // map input model to entity
         var eventEntity = mapper.Map<Event>(im);
         eventEntity.CreatorId = currentUser.User.Id;
-        
+
         // insert event and metadata
         await context.Events.AddAsync(eventEntity, cf);
         // await context.EventMetadataEntries.AddRangeAsync(eventMetadata, cf);
-        
+
         // save changes and map to view model
         await context.SaveChangesAsync(cf);
         var mappedEvent = mapper.Map<EventVM>(eventEntity);
-        
+
         // return created event
         return mappedEvent;
     }
@@ -37,39 +38,41 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
             .Include(@event => @event.Metadata)
             .Include(@event => @event.Creator)
             .FirstOrDefaultAsync(e => e.Id == id, cf);
-        
+
         // if event not found, return null
         if (eventEntity == null)
         {
             return null;
         }
-        
+
         // map to EventVM
         var mappedEvent = mapper.Map<EventVM>(eventEntity);
         mappedEvent.Metadata = eventEntity.Metadata.Select(mapper.Map<EventMetadataVM>);
-        
+
         // return event view model
         return mappedEvent;
     }
 
-    public async Task<IEnumerable<EventVM>> GetCollectionByCreatorAsync(Guid creatorId, List<string> includeMetadataKeys,
+    public async Task<IEnumerable<EventVM>> GetCollectionByCreatorAsync(Guid creatorId,
+        List<string> includeMetadataKeys,
+        PaginationOptions pagination,
         CancellationToken cf = default)
     {
         // retrieve events by creator and include metadata keys
         var query = await context.Events.AsNoTracking()
-            .Include(@event => @event.Creator)
             .Include(@event => @event.Metadata)
+            .Skip((pagination.Page - 1) * pagination.PageSize)
+            .Take(pagination.PageSize)
             .ToListAsync(cf);
 
         // map to EventVM
-        
         return query.Select(e =>
         {
             var eventVm = mapper.Map<EventVM>(e);
-            
+
             eventVm.Metadata = eventVm.Metadata
                 .Where(em => includeMetadataKeys.Contains(em.Key)).ToList();
-            
+
             return eventVm;
         });
     }
@@ -89,7 +92,7 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
     public async Task<BaseCollectionVM<EventVM>> GetAllPaginatedAndFilteredAsync(PaginationOptions paginationOptions,
         EventFilter filter, CancellationToken cf = default)
     {
-        var query = context.Events.AsNoTracking();
+        var query = context.Events.Include(e => e.Metadata).AsNoTracking();
 
         // filter for title and description by the property TextSearch
         if (!string.IsNullOrEmpty(filter.TextSearch))
@@ -98,58 +101,95 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
                                      || e.Description.Contains(filter.TextSearch));
         }
 
-        // pagination stage
-        query = query
-            .Skip((paginationOptions.Page - 1) * paginationOptions.PageSize)
-            .Take(paginationOptions.PageSize);
-        
-        // retrieve events and their count
-        var count = await query.CountAsync(cf);
-        var items = await query.ToListAsync(cf);
-        
-        // join paginated events with their corresponding metadata
-        var metadataQuery = await context.EventMetadataEntries.AsNoTracking().Join(items,
-            m => m.Id,
-            e => e.Id,
-            (m, e) => new { Event = e, MetadataEntry = m })
-            .ToListAsync(cf);
-        
-        // map to EventVM
-        var result = metadataQuery.Select(e => e.Event).Distinct().Select(e =>
+        var parameter = Expression.Parameter(typeof(string), "x");
+        Expression? combinedCondition = null;
+        foreach (var metadataFilter in filter.MetadataFilters)
         {
-            var eventVm = mapper.Map<EventVM>(e);
-            
-            eventVm.Metadata = metadataQuery
-                .Where(em => em.Event.Id == e.Id)
-                .Select(em => em.MetadataEntry)
-                .Select(mapper.Map<EventMetadataVM>);
-            
-            return eventVm;
-        }); 
-        
-        // return paginated and filtered events
-        return new BaseCollectionVM<EventVM>
+            // Build conditions for [key] == 'KeyN' and [value] == 'ValueN'
+            var keyProperty = Expression.Property(parameter, nameof(EventMetadata.Key));
+            var valueProperty = Expression.Property(parameter, nameof(EventMetadata.Value));
+
+            var keyCondition = Expression.Equal(keyProperty, Expression.Constant(metadataFilter.Key));
+            var valueCondition = Expression.Equal(valueProperty, Expression.Constant(metadataFilter.Value));
+
+            var combinedFilter = Expression.AndAlso(keyCondition, valueCondition);
+
+            // Combine with OR
+            combinedCondition = combinedCondition == null
+                ? combinedFilter
+                : Expression.OrElse(combinedCondition, combinedFilter);
+        }
+
+        if (combinedCondition != null)
         {
-            Items = result,
-            Count = count
-        };
+            var events = await query.Join(
+                    context.EventMetadataEntries.Where(
+                        Expression.Lambda<Func<EventMetadata, bool>>(combinedCondition, parameter)),
+                    e => e.Id,
+                    m => m.EventId,
+                    (e, m) => new { Event = e, MetadataEntry = m })
+                .GroupBy(e => e.Event)
+                .Skip((paginationOptions.Page - 1) * paginationOptions.PageSize)
+                .Take(paginationOptions.PageSize)
+                .ToListAsync(cf);
+
+            return new BaseCollectionVM<EventVM>
+            {
+                Items = [],
+                Count = 0
+            };
+        }
+        else
+        {
+            var events = await query
+                .Include(e => e.Metadata)
+                .Skip((paginationOptions.Page - 1) * paginationOptions.PageSize)
+                .Take(paginationOptions.PageSize)
+                .ToListAsync(cf);
+            
+            var count = await query.CountAsync(cf);
+            var items = await query.ToListAsync(cf);
+            
+            // map to EventVM
+            var result = events.Select(e =>
+            {
+                var eventVm = mapper.Map<EventVM>(e);
+
+                eventVm.Metadata = eventVm.Metadata
+                    .Where(em => filter.IncludeMetadataKeys.Contains(em.Key)).ToList();
+
+                return eventVm;
+            });
+
+            // return paginated and filtered events
+            return new BaseCollectionVM<EventVM>
+            {
+                Items = result,
+                Count = count
+            };
+        }
     }
-    
+
     public async Task<bool> DeleteAsync(Guid eventId, CancellationToken cf = default)
     {
         // retrieve event
         var eventEntity = await context.Events.FirstOrDefaultAsync(e => e.Id == eventId, cf);
-        
+
         // if event not found, return null
         if (eventEntity == null)
         {
             return false;
         }
-        
+
+        if (eventEntity.CreatorId != currentUser.User.Id)
+        {
+            return false;
+        }
+
         // remove event and save changes
         context.Events.Remove(eventEntity);
         await context.SaveChangesAsync(cf);
-        
+
         return true;
     }
 
@@ -165,7 +205,13 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
         {
             return null;
         }
-        
+
+        // if event doesn't belong to currentUser return null
+        if (eventEntity.CreatorId != currentUser.User.Id)
+        {
+            return null;
+        }
+
         // check if metadata key already exists
         var existingMetadata = eventEntity.Metadata.FirstOrDefault(m => m.Key == metadata.Key);
         if (existingMetadata != null)
@@ -185,6 +231,7 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
             };
             await context.EventMetadataEntries.AddAsync(newMetadata, cf);
         }
+
         // save changes and return updated event
         await context.SaveChangesAsync(cf);
         return await GetByIdAsync(eventId, cf);
@@ -196,19 +243,25 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
         var eventEntity = await context.Events
             .Include(e => e.Metadata)
             .FirstOrDefaultAsync(e => e.Id == eventId, cf);
-        
+
         // if event not found, return null
         if (eventEntity == null)
         {
             return null;
         }
-        
+
+        // if event doesn't belong to currentUser return null
+        if (eventEntity.CreatorId != currentUser.User.Id)
+        {
+            return null;
+        }
+
         // if metadata key not found, return null
         if (eventEntity.Metadata.FirstOrDefault(m => m.Key == key) == null)
         {
             return null;
         }
-        
+
         // remove metadata and save changes
         context.EventMetadataEntries.Remove(eventEntity.Metadata.FirstOrDefault(m => m.Key == key)!);
         await context.SaveChangesAsync(cf);
@@ -219,11 +272,19 @@ public class EventService(IMapper mapper, ICurrentUser currentUser, ApplicationD
     {
         // retrieve event
         var eventEntity = await context.Events.FirstOrDefaultAsync(e => e.Id == id, cf);
+
+        // if event doesn't exist return false
         if (eventEntity == null)
         {
             return false;
         }
-        
+
+        // if event doesn't belong to currentUser return false
+        if (eventEntity.CreatorId != currentUser.User.Id)
+        {
+            return false;
+        }
+
         // update cancellation status and save changes
         eventEntity.Cancelled = cancelled;
         context.Events.Update(eventEntity);
